feat: add collectiblesListMenu
