+++ packages/desktop/components/menus/TokenListMenu.svelte
@@ -25,14 +25,6 @@
         menu?.close()
     }
 
-    function onImportErc721TokenClick(): void {
-        openPopup({
-            id: PopupId.ImportErc721Token,
-            overflow: true,
-        })
-        menu?.close()
-    }
-
     function refreshTokenMetadata(): void {
         refreshAccountTokensForActiveProfile(true)
         showNotification({
@@ -70,11 +62,6 @@
                 icon: IconName.Import,
                 title: localize('actions.importErc20Token'),
                 onClick: onImportErc20TokenClick,
-            },
-            {
-                icon: IconName.ArrowDown,
-                title: localize('actions.importErc721Token'),
-                onClick: onImportErc721TokenClick,
             },
             {
                 icon: IconName.Refresh2,
