+++ packages/desktop/views/dashboard/collectibles/views/CollectiblesGalleryView.svelte
@@ -5,7 +5,7 @@
     import { Button, IconName, Text, Pill } from '@bloomwalletio/ui'
     import { PopupId, openPopup } from '@desktop/auxiliary/popup'
     import { getActiveNetworkId } from '@core/network'
-    import { EmptyListPlaceholder } from '@components'
+    import { CollectiblesListMenu, EmptyListPlaceholder } from '@components'
 
     function onReceiveClick(): void {
         openPopup({
@@ -26,8 +26,9 @@
                     <Text textColor="secondary">{String($queriedNfts.length ?? '')}</Text>
                 </Pill>
             </div>
-            <div class="flex items-center" style="height: 40px">
+            <div class="flex items-center gap-2" style="height: 40px">
                 <SearchInput bind:value={$nftSearchTerm} />
+                <CollectiblesListMenu />
             </div>
         </div>
